# swagger/Quasar Fire Operation_swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"quasar-fire-controller","description":"Quasar Fire Controller"}],"paths":{"/api/quasarfire/topsecret":{"post":{"tags":["quasar-fire-controller"],"summary":"topSecret","operationId":"topSecretUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"satelites","description":"satelites","required":true,"schema":{"type":"array","items":{"$ref":"#/definitions/Satelite"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TopSecretResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/quasarfire/topsecret_split/{satelite_name}":{"post":{"tags":["quasar-fire-controller"],"summary":"topSecretSplit","operationId":"topSecretSplitUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"requestSplit","description":"requestSplit","required":true,"schema":{"$ref":"#/definitions/TopSecretRequestSplit"}},{"name":"satelite_name","in":"path","description":"satelite_name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TopSecretResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Satelite":{"type":"object","properties":{"distance":{"type":"number","format":"double"},"message":{"type":"array","items":{"type":"string"}},"name":{"type":"string"}},"title":"Satelite"},"ShipPosition":{"type":"object","properties":{"xPosition":{"type":"number","format":"double"},"yPosition":{"type":"number","format":"double"}},"title":"ShipPosition"},"TopSecretRequestSplit":{"type":"object","properties":{"distance":{"type":"number","format":"double"},"message":{"type":"array","items":{"type":"string"}}},"title":"TopSecretRequestSplit"},"TopSecretResponse":{"type":"object","properties":{"message":{"type":"string"},"position":{"$ref":"#/definitions/ShipPosition"}},"title":"TopSecretResponse"}}}